Move files into server/lib and server/service folders
