+++ packages/server-web/src/server/lib/mongo-autoincrement.js
