+++ packages/storage-events/src/event.js
@@ -1,5 +1,8 @@
 const timestamps = require("mongoose-timestamp");
-const { createSequenceSchema, getNextSequenceValue } = require("@todastic/server-web/src/mongo-autoincrement.js");
+const {
+  createSequenceSchema,
+  getNextSequenceValue
+} = require("@todastic/server-web/src/server/lib/mongo-autoincrement.js");
 
 module.exports = { createEventModel };
